Don't import modin until after version check
