+++ python/requirements.txt
@@ -32,7 +32,7 @@ flask
 gym
 lz4
 opencv-python-headless==4.3.0.36
-pandas>=1.0.5
+pandas >= 1.0.5
 scipy==1.4.1
 tabulate
 tensorboardX >= 1.9
